Added option to not run the info-beamer show

+++ run-beamer.sh
@@ -1,4 +1,5 @@
 #!/bin/bash
+[ -f /boot/no-run ] && exit
 [ -f /boot/room ] && cp /boot/room /home/webcamp/info-beamer/webcamp-2015/schedule/room
 nohup /home/webcamp/info-beamer-pi/info-beamer /home/webcamp/info-beamer/webcamp-2015 &
 /home/webcamp/info-beamer/check-git.sh
